pyproject.toml
[tool.poetry]
name = "triage-dashboard"
version = "0.1.0"
description = ""
authors = ["Josh Peak <neozenith.dev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
PyGithub = "^1.55"
dash = "^2.3.0"
pandas = "^1.4.1"
rich = "^12.0.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
invoke-common-tasks = "^0.2.4"
mypy = "^0.941"
pytest = "^7.1.1"
flake8 = "^4.0.1"
black = "^22.1.0"
flake8-docstrings = "^1.6.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
import_heading_stdlib = "Standard Library"
import_heading_firstparty = "Our Libraries"
import_heading_thirdparty = "Third Party"

[tool.mypy]
pretty = true
show_error_codes = true
show_column_numbers = true
show_error_context = true
exclude = [
  'tests/',
  'tasks\.py'
]
follow_imports = 'silent'
ignore_missing_imports = true
# Work your way up to these:
disallow_incomplete_defs = true
# disallow_untyped_defs = true 
# strict = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-s -vvv --color=yes --cov=. --no-cov-on-fail"

[tool.coverage.run]
omit = ["tests/*", "**/__init__.py", "tasks.py"]
branch = true
